Completed the Filetypes split.
Completed splitting filetypes into a different app space.  I need to sit down and examine the code for further simplification, but at this time it's fine.

I will finish the split of the thumbnail creation system first, because that will certainly need a significant rewrite for the split and will most likely start the simplification of the code.